+++ CHANGELOG.md
@@ -10,5 +10,6 @@ Previous version: 0.0.0b1
 - the MorphingOperator interface was implemented 
     - AddAtom operator class was added that adheres to this interface
     - RemoveAtom operator class was added that adheres to this interface
+    - AddBond operator class was added that adheres to this interface
 
 ## Fixes